Reorginized project for compatibility
Renamed pagerank function to GenerateMatrix, and created new function
called PageRank which calls GenerateMatrix to build the adjacency matrix
used in the rest of the algorithm.

PageRank modifies the adjacency matrix to accommodate the damping
factor, then iteratively recalculates a state vector using this
transition matrix (right now the loop for that doesn't work). It appears
that the most efficient way to calculate a new state vector is by matrix
multiplication with the transition matrix and the existing state vector.

We still need to determine when we have reached a steady state, and use
the steady state vector to rank the original entries.